Update revolutionizing-digital-payments.mdx
Added the following at the end of the blog:

"We believe so strongly in Stacks Pay that we are submitting it as a Stacks Improvement Proposal (SIP) to make it the standard everyone uses, harmonizing payment transactions. The revolution in digital payments is here. Are you ready to be part of it? Join us at boom.money."